Added changes to input storage widget based on reviews
