+++ src/modules/dashboard/components/ActionsPage/InputStorageWidget/InputStorageWidget.css.d.ts
@@ -2,3 +2,4 @@ export const widgetForm: string;
 export const learnMoreLink: string;
 export const currentValueText: string;
 export const formFooter: string;
+export const inputStorageSlotContainer: string;
